+++ cuda/dmibulk.cu
@@ -35,7 +35,7 @@ adddmibulk(float* __restrict__ Hx, float* __restrict__ Hy, float* __restrict__ H
            float* __restrict__ Ms_, float Ms_mul,
            float* __restrict__ aLUT2d, float* __restrict__ DLUT2d,
            uint16_t* __restrict__ regions,
-           float cx, float cy, float cz, int Nx, int Ny, int Nz, uint8_t PBC, uint8_t OpenBC) {
+           float cx, float cy, float cz, int Nx, int Ny, int Nz, uint16_t PBC, uint16_t OpenBC) {
 
     int ix = blockIdx.x * blockDim.x + threadIdx.x;
     int iy = blockIdx.y * blockDim.y + threadIdx.y;
