+++ cuda/maxangle.cu
@@ -8,7 +8,7 @@ extern "C" __global__ void
 setmaxangle(float* __restrict__ dst,
             float* __restrict__ mx, float* __restrict__ my, float* __restrict__ mz,
             float* __restrict__ aLUT2d, uint16_t* __restrict__ regions,
-            int Nx, int Ny, int Nz, uint8_t PBC) {
+            int Nx, int Ny, int Nz, uint16_t PBC) {
 
     int ix = blockIdx.x * blockDim.x + threadIdx.x;
     int iy = blockIdx.y * blockDim.y + threadIdx.y;
